lfi-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
arviz=0.11.0=pyhd8ed1ab_0
atk-1.0=2.36.0=h3371d22_4
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
binutils_impl_linux-64=2.36.1=h193b22a_2
binutils_linux-64=2.36=hf3e587d_33
bokeh=2.4.1=py38h578d9bd_1
botorch=0.4.0=pyhd8ed1ab_0
bzip2=1.0.8=h7f98852_4
c-ares=1.17.2=h7f98852_0
ca-certificates=2021.10.8=ha878542_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cairo=1.16.0=h6cf1ce9_1008
certifi=2021.10.8=py38h578d9bd_0
cffi=1.14.6=py38h3931269_1
cftime=1.5.1=py38h6c62de6_0
click=7.1.2=pyh9f0ad1d_0
cloudpickle=1.6.0=py_0
colorama=0.4.4=pyh9f0ad1d_0
curl=7.79.1=h494985f_1
cycler=0.10.0=py_2
cytoolz=0.11.0=py38h497a2fe_3
dask=2.30.0=py_0
dask-core=2.30.0=py_0
decorator=5.1.0=pyhd8ed1ab_0
distributed=2.30.1=py38h578d9bd_0
elfi=0.8.2=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.4.1=h9c3ff4c_0
fastprogress=1.0.0=py_0
ffmpeg=4.4.0=hca11adc_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.13.1=hba837de_1005
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.10.4=h0708190_1
fribidi=1.0.10=h36c2ea0_0
fsspec=2021.10.1=pyhd8ed1ab_0
future=0.18.2=py38h578d9bd_3
gcc_impl_linux-64=7.5.0=hda68d29_13
gcc_linux-64=7.5.0=h47867f9_33
gdk-pixbuf=2.42.6=h04a7f16_0
gettext=0.19.8.1=h73d1719_1008
giflib=5.2.1=h36c2ea0_2
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
gpy=1.10.0=py38h6c62de6_1
gpytorch=1.5.1=pyhd8ed1ab_0
graphite2=1.3.13=h58526e2_1001
graphviz=2.49.1=h85b4f2f_0
gtk2=2.24.33=h539f30e_1
gts=0.7.6=h64030ff_2
gxx_impl_linux-64=7.5.0=h64c220c_13
gxx_linux-64=7.5.0=h555fc39_33
gym=0.19.0=py38he5a9106_0
h5py=3.4.0=nompi_py38hfbb2109_101
harfbuzz=3.0.0=h83ec7ef_1
hdbscan=0.8.27=py38h5c078b8_0
hdf4=4.2.15=h10796ff_3
hdf5=1.12.1=nompi_h7f166f4_101
heapdict=1.0.1=py_0
icu=68.1=h58526e2_0
ipykernel=5.5.5=py38hd0cf306_0
ipyparallel=7.1.0=pyhd8ed1ab_0
ipython=7.28.0=py38he5a9106_0
ipython_genutils=0.2.0=py_1
jbig=2.1=h7f98852_2003
jedi=0.18.0=py38h578d9bd_2
jinja2=3.0.2=pyhd8ed1ab_0
joblib=1.1.0=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.8.1=py38h578d9bd_0
kernel-headers_linux-64=2.6.32=he073ed8_14
kiwisolver=1.3.2=py38h1fd1430_0
krb5=1.19.2=h48eae69_2
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
lerc=3.0=h9c3ff4c_0
libblas=3.9.0=12_linux64_mkl
libcblas=3.9.0=12_linux64_mkl
libcurl=7.79.1=h494985f_1
libdeflate=1.8=h7f98852_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.4.2=h9c3ff4c_4
libgcc-ng=11.2.0=h1d223b6_11
libgd=2.3.3=h6ad9fb6_0
libgfortran-ng=11.2.0=h69a702a_11
libgfortran5=11.2.0=h5c6108e_11
libglib=2.70.0=h174f98d_0
libgomp=11.2.0=h1d223b6_11
libgpuarray=0.7.6=h7f98852_1003
libiconv=1.16=h516909a_0
liblapack=3.9.0=12_linux64_mkl
libllvm11=11.1.0=hf817b99_2
libnetcdf=4.8.1=nompi_hb3fd0d9_101
libnghttp2=1.43.0=ha19adfc_1
libnsl=2.0.0=h7f98852_0
libpng=1.6.37=h21135ba_2
libprotobuf=3.18.1=h780b84a_0
librsvg=2.52.2=hc3c00ef_0
libsodium=1.0.18=h36c2ea0_1
libssh2=1.10.0=ha35d2d1_2
libstdcxx-ng=11.2.0=he4da1e4_11
libtiff=4.3.0=h6f004c6_2
libtool=2.4.6=h9c3ff4c_1008
libuuid=2.32.1=h7f98852_1000
libwebp=1.2.1=h3452ae3_0
libwebp-base=1.2.1=h7f98852_0
libxcb=1.13=h7f98852_1003
libxml2=2.9.12=h72842e0_0
libzip=1.8.0=h1c5bbd1_1
libzlib=1.2.11=h36c2ea0_1013
llvm-openmp=12.0.1=h4bd325d_1
llvmlite=0.37.0=py38h4630a5e_0
locket=0.2.0=py_2
lz4-c=1.9.3=h9c3ff4c_1
mako=1.1.5=pyhd8ed1ab_0
markupsafe=2.0.1=py38h497a2fe_0
matplotlib-base=3.4.3=py38hf4fb855_1
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mkl=2021.4.0=h8d4b97c_729
mkl-service=2.4.0=py38h497a2fe_0
msgpack-python=1.0.2=py38h1fd1430_1
ncurses=6.2=h58526e2_4
nest-asyncio=1.5.1=pyhd8ed1ab_0
netcdf4=1.5.7=nompi_py38h2823cc8_103
nettle=3.6=he412f7d_0
networkx=2.5=py_0
ninja=1.10.2=h4bd325d_1
numba=0.54.1=py38h4bf6c61_0
numpy=1.20.3=py38h9894fe3_1
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=3.0.0=h7f98852_1
packaging=21.0=pyhd8ed1ab_0
pandas=1.3.4=py38h43a58ef_0
pango=1.48.10=h54213e6_2
paramz=0.9.5=py_0
parso=0.8.2=pyhd8ed1ab_0
partd=1.2.0=pyhd8ed1ab_0
patsy=0.5.2=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pip=21.3=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
prompt-toolkit=3.0.20=pyha770c72_0
psutil=5.8.0=py38h497a2fe_1
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pycparser=2.20=pyh9f0ad1d_2
pyglet=1.5.16=py38h578d9bd_0
pygments=2.10.0=pyhd8ed1ab_0
pygpu=0.7.6=py38h5c078b8_1002
pymc3=3.9.3=py_1
pynndescent=0.5.5=pyh6c4a22f_0
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.8.12=hf930737_2_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2021.3=pyhd8ed1ab_0
pyyaml=6.0=py38h497a2fe_0
pyzmq=22.3.0=py38h2035c66_0
readline=8.1=h46c0cb4_0
scikit-learn=1.0=py38hacb3eff_1
scipy=1.7.1=py38h56a6a73_0
setuptools=58.2.0=py38h578d9bd_0
six=1.16.0=pyh6c4a22f_0
sleef=3.5.1=h9b69904_2
sortedcontainers=2.4.0=pyhd8ed1ab_0
sqlite=3.36.0=h9cd32fc_2
sysroot_linux-64=2.12=he073ed8_14
tbb=2021.3.0=h4bd325d_0
tblib=1.7.0=pyhd8ed1ab_0
theano=1.0.5=py38hadf7658_1
threadpoolctl=3.0.0=pyh8a188c0_0
tk=8.6.11=h27826a3_1
toolz=0.11.1=py_0
tornado=6.1=py38h497a2fe_1
tqdm=4.62.3=pyhd8ed1ab_0
traitlets=5.1.0=pyhd8ed1ab_0
typing-extensions=3.10.0.2=hd8ed1ab_0
typing_extensions=3.10.0.2=pyha770c72_0
tzdata=2021d=he74cb21_0
umap-learn=0.5.1=py38h578d9bd_1
wcwidth=0.2.5=pyh9f0ad1d_2
wheel=0.37.0=pyhd8ed1ab_1
x264=1!161.3030=h7f98852_1
xarray=0.19.0=pyhd8ed1ab_1
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
zeromq=4.3.4=h9c3ff4c_1
zict=2.0.0=py_0
zlib=1.2.11=h36c2ea0_1013
zstd=1.5.0=ha95c52a_0
